create runall methods
